IMPALA-4652: Add crcutil to toolchain
Note: crcutil does not have any release tarballs, so we borrow the one
built for Apache Kudu. As a result, the build version is a git hash,
rather than e.g. 1.0.

The package builds on all supported platforms, except for CentOS 5.X due
to the outdated autotools on that platform. As a workaround, the script
will download and locally install autoconf / automake / libtool if it
detects it is on CentOS 5.X. Doing so only adds a few seconds to the build.

Change-Id: Ibf8b39082914b1b2932b4ce7efbd2cc4f5f69743